+++ .gitignore
@@ -5,3 +5,4 @@ __pycache__/
 *.o
 *.out
 *.c
+*.egg-info/
