+++ README.md
@@ -5,7 +5,6 @@
 - [ ] functions
 - [ ] type inference
 
-
 ## Compiling
 
 ```
@@ -13,4 +12,13 @@ python3 compile.py example.cl
 llc example.ll -march=x86-64 -o example.s
 gcc -c example.s -o example.o
 gcc example.o -o example.out
+```
+
+## Run tests
+
+```
+pip install -e .[test]
+pytest
+flake8
+mypy cheer
 ```
