+++ compile.sh
@@ -1,4 +1,4 @@
-python3 compile.py -i $1 -o example.ll
+python3 cheer/compile.py -i $1 -o example.ll
 llc example.ll -march=x86-64 -o example.s
 gcc -c example.s -o example.o
 gcc example.o -o a.out
